extract actions to a separate module
