+++ src/components/record/record-new.tsx
@@ -1,18 +1,5 @@
 import React, { useState } from 'react'
-import { ActionFunctionArgs, Form, redirect } from 'react-router-dom'
-import { loadConfig } from '../../services/config-storage'
-import { createRecord } from '../../services/journal-repository'
-
-export async function newRecordAction(
-  args: ActionFunctionArgs
-): Promise<Response> {
-  const values = Object.fromEntries(await args.request.formData())
-
-  const config = loadConfig()
-  await createRecord(values.content.toString(), config)
-
-  return redirect('/')
-}
+import { Form } from 'react-router-dom'
 
 const RecordNew: React.FunctionComponent = () => {
   const [content, setContent] = useState('')
