+++ src/index.tsx
@@ -6,9 +6,10 @@ import reportWebVitals from './reportWebVitals'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 import Layout from './components/layout'
 import RecordsList from './components/record/records-list'
-import RecordNew, { newRecordAction } from './components/record/record-new'
+import RecordNew from './components/record/record-new'
 import RecordView from './components/record/record-view'
 import ConfigGithub from './components/config/config-github'
+import { newRecordAction } from './components/record/record-actions'
 
 const router = createBrowserRouter([
   {
